pyproject.toml
[tool.poetry]
name = "openai-test"
version = "0.1.0"
description = ""
authors = ["Jungminchae <minchae3618@gmail.com>"]
readme = "README.md"
packages = [{include = "openai_test"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.8"
fastapi = "^0.100.0"
uvicorn = "^0.23.0"
sqlalchemy = "^2.0.19"
python-decouple = "^3.8"
pydantic-settings = "^2.0.2"
attrs = "^23.1.0"
supabase = "^1.0.4"
pytz = "^2023.3.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
